+++ CHANGELOG.rst
@@ -2,6 +2,8 @@
 -----
 
 - Adds tile extractor for numpy array and torch tensors.
+- Fixes CheckpointingCallback: no longer delete end of epoch checkpoints by
+  default according to `max_files`.
 
 0.0.9
 -----